+++ CHANGELOG.md
@@ -4,6 +4,7 @@
  * Add localize\_with\_fallback (see initializer).
  * Updated german locale to latest version from rails-i18n.
  * Use mysql2 gem instead of mysql.
+ * Use customized template for scaffold generator.
 
 ## 1.2.3 (2012-02-10)
  * Updated Modernizr to v2.5.2.